Moving a bunch of pixels across the screen
If you have a drawing on the screen and want to move it off the screen.
Or, if you delete the clearslate the pixels can leave a trail if that’s
what you want. Thanks Elana for helping me compress the code and edit
it! Hope this code can help someone.